requirements.txt
python-dotenv==1.0.1
pymongo==4.8.0
requests==2.32.3
tweepy==4.14.0
spacy==3.7.4
en-core-web-sm==3.7.1
vaderSentiment==3.3.2
pdfplumber==0.11.4
pikepdf==9.2.2
openai==1.50.2
tenacity==8.5.0

